Corrected a logic error in the csv column filter. Filter may not support keeping the last column - a comma might be after the last value in each line.

+++ src/corp/Corp/Corp.Services.DataContracts/FloodingAlertWorkflowResponse.cs
@@ -7,7 +7,7 @@ namespace Corp.Services.DataContracts
     public class FloodingAlertWorkflowResponse
     {
         [DataMember(Order = 1)]
-        public double WaterLevel { get; set; }
+        public int WaterLevel { get; set; } // cm
 
         [DataMember(Order = 2)]
         public string WindSpeed { get; set; }
